+++ backend/canisters/community/c2c_client/src/lib.rs
@@ -3,6 +3,7 @@ use community_canister::*;
 
 // Queries
 generate_c2c_call!(c2c_bot_create_channel);
+generate_c2c_call!(c2c_bot_delete_channel);
 generate_c2c_call!(c2c_bot_send_message);
 generate_c2c_call!(c2c_can_issue_access_token_for_channel);
 generate_c2c_call!(c2c_can_issue_access_token);