+++ backend/canisters/local_user_index/impl/src/updates/bot_create_channel.rs
@@ -10,7 +10,7 @@ async fn bot_create_channel(args: Args) -> Response {
 
     let context = match mutate_state(|state| extract_access_context(&args.auth_token, state)) {
         Ok(context) => context,
-        Err(error) => return InvalidRequest(error),
+        Err(error) => return NotAuthenticated(error),
     };
 
     let BotActionScope::Community(details) = context.scope else {