+++ backend/canisters/local_user_index/impl/src/updates/bot_send_message.rs
@@ -13,7 +13,7 @@ async fn bot_send_message(args: Args) -> Response {
 
     let context = match mutate_state(|state| extract_message_access_context(&args, state)) {
         Ok(context) => context,
-        Err(error) => return InvalidRequest(error),
+        Err(response) => return response,
     };
 
     match context.chat {
@@ -59,14 +59,18 @@ struct MessageAccessContext {
     message_id: MessageId,
 }
 
-fn extract_message_access_context(args: &Args, state: &mut RuntimeState) -> Result<MessageAccessContext, String> {
-    let context = extract_access_context(&args.auth_token, state)?;
+fn extract_message_access_context(args: &Args, state: &mut RuntimeState) -> Result<MessageAccessContext, Response> {
+    use Response::*;
+
+    let context = extract_access_context(&args.auth_token, state).map_err(NotAuthenticated)?;
 
     let (chat, thread, message_id) = match context.scope {
         BotActionScope::Chat(details) => {
             if let Some(message_id) = args.message_id {
                 if matches!(context.initiator, BotInitiator::Command(_)) && message_id != details.message_id {
-                    return Err("Message id is already specified in the command access token".to_string());
+                    return Err(InvalidRequest(
+                        "Message id is already specified in the command access token".to_string(),
+                    ));
                 }
             }
             let message_id = args.message_id.unwrap_or(details.message_id);
@@ -74,7 +78,7 @@ fn extract_message_access_context(args: &Args, state: &mut RuntimeState) -> Resu
         }
         BotActionScope::Community(details) => {
             let Some(channel_id) = args.channel_id else {
-                return Err("Channel must be specified for community scope".to_string());
+                return Err(InvalidRequest("Channel must be specified for community scope".to_string()));
             };
             let chat = Chat::Channel(details.community_id, channel_id);
             let message_id = args.message_id.unwrap_or_else(|| state.env.rng().gen::<u64>().into());